server/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1